Ajout de DofusProtocolBuilder Mise à jour du protocol -> 2.8.0

+++ trunk/Protocol/Types/ProtocolTypeManager.cs
@@ -25,7 +25,7 @@ namespace BiM.Protocol.Types
                 if (field != null)
                 {
                     // le cast uint est obligatoire car l'objet n'a pas de type
-                    short id = (short)((uint)field.GetValue(type));
+                    short id = (short)(field.GetValue(type));
 
                     m_types.Add(id, type);
